backend/requirements.txt
Flask==3.0.0
mongoengine==0.28.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.4.0
pymongo==4.6.1
dnspython==2.8.0
certifi==2025.10.5

